+++ app/src/main/java/com/rama/mako/receivers/ScreenLockAdminReceiver.kt
@@ -1,5 +1,0 @@
-package com.rama.mako.receivers
-
-import android.app.admin.DeviceAdminReceiver
-
-class ScreenLockAdminReceiver : DeviceAdminReceiver()